Moving exceptions to brand lib directory
