+++ lib/python/brand/__init__.py
@@ -3,3 +3,5 @@ from .tools import (get_node_parameter_value, get_parameter_value,
                          get_redis_info, main, get_node_io, unpack_string,
                          node_stage)
 from .node import BRANDNode 
+
+from .exceptions import (GraphError, RedisError)
